fix: Discover resources if no organization is configured for the provider (#62)
This fixes the case where no `organization` argument is passed to the
`tfe` provider block.

Before this fix, users with the following configuration would see the
subsequent error message:

```hcl
provider "tfe" {}
```

```shell
│ Error: no organization was specified on the resource or provider
│
│   with module.discovery.data.tfe_variable_set.this["<VARIABLE_SET_NAME>"],
│   on .terraform/modules/discovery/discovery.tf line 11, in data "tfe_variable_set" "this":
│   11: data "tfe_variable_set" "this" {
```

This requires users to supply the `organization` argument to the
provider which isn't robust for a good user experience.