feat: Add tfe_workspace discovery (#68)
Adding the `tfe_workspace` output which is a map of `workspace_ids` with
their configuration that can be iterated over to import workspaces.

This output isn't as generic as I would like it, for example you would
need to conditionally set the resource configuration based on the
returned configuration (`vcs_repo` could be populated or not). An
additional output will be added in the future to make this easier for
consumers.